runtime, ssa: port cosmo and wasm code to go1.27 internal APIs
decoderune now takes and returns a uint index, and the ssa generator's
regMask is a struct rather than a scalar. opGen.go and rewriteWasm.go are
regenerated from the merged _gen sources.

+++ src/runtime/os_cosmo_nt_path.go
@@ -57,7 +57,7 @@ import "unsafe"
 // dst, flipping '/' to '\' when flipSlash is set. Invalid UTF-8 bytes
 // become U+FFFD, matching unicode/utf16's encoder.
 func ntUTF16Append(dst []uint16, s string, flipSlash bool) []uint16 {
-	for i := 0; i < len(s); {
+	for i := uint(0); i < uint(len(s)); {
 		c := s[i]
 		if c < 0x80 {
 			if c == '/' && flipSlash {
